add auth and one-to-many

+++ app/Http/Controllers/TasksController.php
@@ -16,13 +16,24 @@ class TasksController extends Controller
      //getでtasksにアクセスされた場合の「一覧表示」
     public function index()
     {
-        //タスク一覧を取得
-        $tasks = Task::all();
+
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            
+            //indexビューで表示
+            return view('tasks.index', $data);
+        }
         
-        //タスク一覧ビューでそれを表示
-        return view("tasks.index",[
-            "tasks"=>$tasks,
-        ]);
+        //dashboardビューを表示
+        return view("dashboard");
         
     }
 
@@ -34,12 +45,14 @@ class TasksController extends Controller
      //getでtasks/createにアクセスした場合の「タスク追加画面表示処理」
     public function create()
     {
+
         $task = new Task;
         
         //タスク追加ビューを表示
         return view("tasks.create",[
             "task" => $task,
         ]);
+
     }
 
     /**
@@ -51,20 +64,24 @@ class TasksController extends Controller
      //postでtasksにアクセスされた場合の「タスク追加処理」
     public function store(Request $request)
     {
+        
+        $user = \Auth::user();
+        
         //バリデーション
         $request->validate([
             "content" => "required",
             "status" => "required|max:10",
         ]);
         
-        //タスクを作成して保存
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
         
         //トップページへリダイレクトさせる
         return redirect("/");
+
     }
 
     /**
@@ -76,13 +93,21 @@ class TasksController extends Controller
      //getでtasks/idにアクセスされた場合の「取得表示処理」
     public function show($id)
     {
-        //idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
+
+        // idの値で投稿を検索して取得
+        $task = \App\Models\Task::findOrFail($id);
         
-        //タスク詳細ビューで表示
-        return view("tasks.show",[
-            "task" => $task,
-        ]);
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は投稿を表示
+        if (\Auth::id() === $task->user_id) {
+            //タスク詳細ビューで表示
+            return view("tasks.show",[
+                "task" => $task,
+            ]);
+        }
+
+        //トップページへリダイレクトさせる
+        return redirect("/");
+        
     }
 
     /**
@@ -94,13 +119,18 @@ class TasksController extends Controller
      //getでtasks/id/editにアクセスされた場合の「更新画面表示処理」
     public function edit($id)
     {
+
         //idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         
-        //タスク編集ビューで表示
-        return view("tasks.edit",[
-            "task"=>$task,
-        ]);
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は更新画面を表示
+        if (\Auth::id() === $task->user_id) {
+            //タスク編集ビューで表示
+            return view("tasks.edit",[
+                "task"=>$task,
+            ]);
+        }
+
     }
 
     /**
@@ -113,6 +143,7 @@ class TasksController extends Controller
      //putまたはpatchでtasks/idにアクセスされた場合の「更新処理」
     public function update(Request $request, $id)
     {
+        
         //バリデーション
         $request->validate([
             "content" => "required",
@@ -122,11 +153,14 @@ class TasksController extends Controller
         //idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         
-        //タスク更新
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
-        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は投稿を更新
+        if (\Auth::id() === $task->user_id) {
+            //タスク更新
+            $task->content = $request->content;
+            $task->status = $request->status;
+            $task->save();
+        }
+
         //トップページへリダイレクトさせる
         return redirect("/");
     }
@@ -140,13 +174,16 @@ class TasksController extends Controller
      //deleteでtasks/idにアクセスされた場合の「削除処理」
     public function destroy($id)
     {
-        //idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
+
+        // idの値で投稿を検索して取得
+        $task = \App\Models\Task::findOrFail($id);
         
-        //タスクを削除
-        $task->delete();
-        
-        //トップページへリダイレクトさせる
-        return redirect("/");
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            //トップページへリダイレクトさせる
+            return redirect("/");
+        }
+
     }
 }
